% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ae.R
\name{ae}
\alias{ae}
\title{Analisis exploratorio de los datos}
\usage{
ae(datos)
}
\arguments{
\item{datos}{La base de datos que se creo utilizando la
funcion inci.}
}
\description{
La funcion lleva a cabo un pequenho analisis
exploratorio de los datos. Calcula el numero total de
incidencias, las incidencias por tipo, las incidencias
por mes, y por tipo y por mes, tanto para las incidencias
provocadas en regimen de enfriamiento como de calentamiento.
}
\examples{
library(incidents)
## Tarda unos minutos
# misdatos <- readbbdd(file = "lista_all.txt", file2 = "lista.txt)
# misdatos2<-inci(misdatos)
# misdatos2 <-misdatos2[,-c(7,8)]
# mianalisis <- ae(misdatos2)
}
\author{
Nora M. Villanueva y Javier Roca Pardinas
}

