% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hour2day.R
\name{hour2day}
\alias{hour2day}
\title{Creación de la base de datos diaria}
\usage{
hour2day(datos)
}
\arguments{
\item{datos}{La base de datos creada con la función inci.}
}
\description{
La función crea una nueva base de datos diaria, es
decir, resume la información horaria en información diaria.
Para ello se han creado nuevas variables. Por ejemplo, temperatura
ambiente mínima, máxima o media.
}
\examples{
library(incidents)
## Ojo, tarda unos minutos
# misdatos <- readbbdd(file = "lista_all.txt", file2 = "lista.txt)
# misdatos2<-inci(misdatos)
# datosdia <- hour2day(misdatos2)
## Elimino las variables de impulsion y retorno tanto de agua,
## como de aire.
# datosdia <- datosdia[,c(1:22, 32,33)]
}
\author{
Nora M. Villanueva y Javier Roca Pardinas
}

