% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inci.R
\name{inci}
\alias{inci}
\title{Creación de la variable Inci1 e Inci2}
\usage{
inci(datos)
}
\arguments{
\item{datos}{La base de datos que se leyó y unificó utilizando la
función readbbdd.}
}
\description{
La función permite crear las variables respuesta
inci1 (incidencias en régimen de enfriamiento) e inci2 (inci-
dencias en régimen de calentamiento).

Régimen enfriamiento: existe incidencia si
Temperatura ambiente > Temperatura consigna + 2 C, siendo la
temperatura de consigna = 22 C, si la temperatura de impulsión
del aire es menor que la temperatura ambiente − 5 C.

Régimen calentamiento: existe incidencia si
Temperatura ambiente < Temperatura consigna - 2 C, siendo la
temperatura de consigna = 22 C, si la temperatura de impulsión
del aire es mayor que la temperatura ambiente + 5 C.
}
\examples{
library(incidents)
# Ojo, tarda unos minutos
# misdatos <- readbbdd(file = "lista_all.txt", file2 = "lista.txt)
# misdatos2<-inci(misdatos)
# elimino las variables: impulsion.agua.fria e impulsion.agua.caliente
# misdatos2 <-misdatos2[,-c(7,8)]
}
\author{
Nora M. Villanueva y Javier Roca Pardinas
}

