% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelpred.R
\name{modelpred}
\alias{modelpred}
\title{Construcción de los modelos de predicción de incidencias}
\usage{
modelpred(data, newdata, pred.days = 1)
}
\arguments{
\item{data}{La base de datos creada con la función hour2day.}

\item{newdata}{Los nuevos valores a predecir incidencias.
Nota: tiene que ser un objeto de la clase data.frame.}

\item{pred.days}{Por defecto es 1. Indica que la predicción se
hará para el día siguiente. Por ejemplo, si se pretende hacer
predicción a dos días, se sustituye el 1 por el dos.}
}
\description{
La función construye dos modelos de predicción
de incidencias. Uno para incidencias provocadas en régimen de
enfriamiento y el otro para aquellas provocadas  en régimen
de calentamiento. Los modelos incluyen las covariables
seleccionadas:
amb.max: es la temperatura ambiente máxima (diaria) antes de que abra
         la tienda, es decir,el mismo día que se quiere hacer la
         predicción.
amb.mean.on: es la temperatura ambiente media (diaria) durante el horario
         comercial, es decir, el día antes de hacer la predicción.
cli.mean: consumos de aire acondicionado (diario) durante el horario
         comercial, es decir, el día antes de hacer la predicción.
inci.5: número de incidencias los 5 días antes de hacer la predicción.
       Es decir, si el día anterior hubo incidencia, dicha variable
       sería igual a 1, si los 5 días antes ha habido incidencia, sería
       igual a 5. Si ha habido incidencia durante 10 días consecutivos,
       el valor máximo de dicha variable es de 5.
}
\examples{
library(incidents)
# Ojo, tarda unos minutos
# misdatos <- readbbdd(file = "lista_all.txt", file2 = "lista.txt)
# misdatos2<-inci(misdatos)
# misdatos2 <-misdatos2[,-c(7,8)]
# datosdia <- hour2day(misdatos2)
## Elimino las variables de impulsion y retorno tanto de agua,
## como de aire.
# datosdia <- datosdia[,c(1:22, 32,33)]
## Por ejemplo: queremos predecir si va a haber incidencia
## en una tienda con los siguientes valores:
# datos2pred <-data.frame (amb.max = 40, amb.mean.on = 5,
#              cli.mean = 150, inci.5d = 3)
# modelpred(data = datosdia, pred.days = 1, newdata = datos2pred)
}
\author{
Nora M. Villanueva y Javier Roca Pardinas
}

