% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inci.R
\name{inci}
\alias{inci}
\title{Creacion de la variable Inci1 e Inci2}
\usage{
inci(datos)
}
\arguments{
\item{datos}{La base de datos que se leyo y unifico utilizando la
funcion readbbdd.}
}
\description{
La funcion permite crear las variables respuesta
inci1, incidencias en regimen de enfriamiento, e inci2,
incidencias en regimen de calentamiento.

Regimen enfriamiento, existe incidencia si
Temperatura ambiente mayor que  Temperatura consigna mas  2 C,
siendo la temperatura de consigna igual a  22 C, si la
temperatura de impulsion del aire es menor que la
temperatura ambiente menos 5 C.

Regimen calentamiento, existe incidencia si
Temperatura ambiente menor que la  Temperatura consigna
menos 2 C, siendo la temperatura de consigna
igual a 22 C, si la temperatura de impulsion del aire
 es mayor que la  temperatura ambiente mas 5 C.
}
\examples{
library(incidents)
## Tarda unos minutos
# misdatos <- readbbdd(file = "lista_all.txt", file2 = "lista.txt)
# misdatos2<-inci(misdatos)
}
\author{
Nora M. Villanueva y Javier Roca Pardinas
}

